+++ lib/rounded_background_text.dart
@@ -1,5 +1,6 @@
 library rounded_background_text;
 
-export 'src/rounded_background_text_field.dart';
-export 'src/rounded_background_text_span.dart';
 export 'src/rounded_background_text.dart' hide foregroundColor;
+export 'src/rounded_background_text_field.dart';
+export 'src/rounded_background_text_selectable.dart';
+export 'src/rounded_background_text_span.dart';